requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
anthropic>=0.29.0
openai>=1.30.0
pydantic>=2.7.0
python-dateutil>=2.8.2
dateparser>=1.2.0
pytz>=2024.1
